+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/Dialogs/ResolveContextualInfoResponses.de.json
