+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/Dialogs/UserInfoOptions.cs
@@ -1,7 +1,7 @@
-﻿using Microsoft.Bot.Builder.Skills.Contextual.Models;
+﻿using Microsoft.Bot.Builder.Solutions.Contextual.Models;
 using System.Collections.Generic;
 
-namespace Microsoft.Bot.Builder.Skills.Contextual.Dialogs
+namespace Microsoft.Bot.Builder.Solutions.Contextual.Dialogs
 {
     public class UserInfoOptions
     {
