+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/Models/Relationship/PossessivePronoun.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Microsoft.Bot.Builder.Skills.Contextual.Models
+﻿namespace Microsoft.Bot.Builder.Solutions.Contextual.Models
 {
     public class PossessivePronoun
     {
