+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/Models/Relationship/RelatedEntityInfo.cs
@@ -1,4 +1,4 @@
-﻿namespace Microsoft.Bot.Builder.Skills.Contextual.Models
+﻿namespace Microsoft.Bot.Builder.Solutions.Contextual.Models
 {
     public class RelatedEntityInfo
     {
