+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/Models/UserHistoryInfo.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Microsoft.Bot.Builder.Skills.Contextual.Models
+namespace Microsoft.Bot.Builder.Solutions.Contextual.Models
 {
     public class UserHistoryInfo
     {
