+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/Models/UserInfoState.cs
@@ -1,6 +1,6 @@
-﻿namespace Microsoft.Bot.Builder.Skills.Contextual.Models
+﻿namespace Microsoft.Bot.Builder.Solutions.Contextual.Models
 {
-    using Microsoft.Bot.Builder.Skills.Contextual.Resources;
+    using Microsoft.Bot.Builder.Solutions.Contextual.Resources;
     using Microsoft.Bot.Builder.Solutions.Resources;
     using System.Collections.Generic;
     using System.Linq;
