+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/Resources/RelationshipStrings.es.resx
