+++ lib/csharp/microsoft.bot.builder.solutions/microsoft.bot.builder.solutions/Contextual/UserStateContextResolver.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Microsoft.Bot.Builder.Skills.Contextual.Models;
-using Microsoft.Bot.Builder.Skills.Contextual.Resources;
+using Microsoft.Bot.Builder.Solutions.Contextual;
+using Microsoft.Bot.Builder.Solutions.Contextual.Models;
 
-namespace Microsoft.Bot.Builder.Skills.Contextual
+namespace Microsoft.Bot.Builder.Solutions.Contextual
 {
     public class UserStateContextResolver : IContextResolver
     {
